Move slice reify and unify to core
